# [6.0] Fix update pipeline perf regression