https://github.com/dotnet/efcore/pull/25952 decreased throughput in mixed update scenario by about 20-40%. These changes mostly remove this degradation:
- When sorting update commands add Deleted + Inserted number of edges instead of Deleted * Inserted
- Don't allocate a collection for edges if there's only one edge between given two vertices
- Reuse the `Multigraph` instance
- Change the benchmark to reuse the context instance to simulate pooling behavior

<details>
<summary>Before #25952</summary>

|         Method | Async | Batching |     Mean |   Error |  StdDev |  Op/s |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |------ |--------- |--------- |-------- |-------- |------ |---------- |------ |------ |---------- |
| UpdatePipeline | False |    False | 122.4 ms | 1.78 ms | 1.49 ms | 8.169 | 1000.0000 |     - |     - |  11.31 MB |
| UpdatePipeline | False |     True | 227.3 ms | 4.48 ms | 5.16 ms | 4.400 | 2000.0000 |     - |     - |  15.33 MB |
| UpdatePipeline |  True |    False | 129.5 ms | 1.35 ms | 1.20 ms | 7.725 | 1000.0000 |     - |     - |  11.53 MB |
| UpdatePipeline |  True |     True | 243.4 ms | 1.13 ms | 0.94 ms | 4.109 | 2000.0000 |     - |     - |  17.51 MB |

</details>


<details>
<summary>After #25952</summary>

|         Method | Async | Batching |     Mean |   Error |   StdDev |  Op/s |     Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|--------------- |------ |--------- |--------- |-------- |--------- |------ |---------- |---------- |---------- |---------- |
| UpdatePipeline | False |    False | 201.0 ms | 3.80 ms |  5.08 ms | 4.975 | 5000.0000 | 3000.0000 | 1000.0000 |  32.94 MB |
| UpdatePipeline | False |     True | 279.2 ms | 4.44 ms |  3.93 ms | 3.581 | 5000.0000 | 3000.0000 | 1000.0000 |  36.96 MB |
| UpdatePipeline |  True |    False | 196.6 ms | 3.72 ms |  4.43 ms | 5.087 | 5000.0000 | 3000.0000 | 1000.0000 |  33.15 MB |
| UpdatePipeline |  True |     True | 370.5 ms | 7.09 ms | 12.79 ms | 2.699 | 6000.0000 | 4000.0000 | 1000.0000 |  39.16 MB |

</details>


<details>
<summary>After this fix</summary>

|         Method | Async | Batching |     Mean |   Error |  StdDev |  Op/s |     Gen 0 |     Gen 1 | Gen 2 | Allocated |
|--------------- |------ |--------- |--------- |-------- |-------- |------ |---------- |---------- |------ |---------- |
| UpdatePipeline | False |    False | 125.8 ms | 2.51 ms | 6.44 ms | 7.950 | 1000.0000 |         - |     - |  11.08 MB |
| UpdatePipeline | False |     True | 226.3 ms | 3.06 ms | 2.71 ms | 4.420 | 2000.0000 | 1000.0000 |     - |  15.11 MB |
| UpdatePipeline |  True |    False | 126.8 ms | 2.42 ms | 2.69 ms | 7.883 | 1000.0000 |         - |     - |  11.29 MB |
| UpdatePipeline |  True |     True | 277.9 ms | 5.54 ms | 5.93 ms | 3.599 | 2000.0000 | 1000.0000 |     - |  17.29 MB |

</details>