@AndriySvyryd 2021-09-17
Turning pooling off doesn't have a significant effect on the results

|         Method | Async | Batching |     Mean |   Error |  StdDev |  Op/s |     Gen 0 |     Gen 1 | Gen 2 | Allocated |
|--------------- |------ |--------- |--------- |-------- |-------- |------ |---------- |---------- |------ |---------- |
| UpdatePipeline | False |    False | 135.2 ms | 2.58 ms | 3.07 ms | 7.399 | 1000.0000 |         - |     - |  11.19 MB |
| UpdatePipeline | False |     True | 205.2 ms | 4.07 ms | 3.61 ms | 4.873 | 2000.0000 | 1000.0000 |     - |  15.22 MB |
| UpdatePipeline |  True |    False | 129.7 ms | 2.56 ms | 3.50 ms | 7.713 | 1000.0000 |         - |     - |   11.4 MB |
| UpdatePipeline |  True |     True | 243.0 ms | 3.73 ms | 3.49 ms | 4.116 | 2000.0000 | 1000.0000 |     - |   17.4 MB |