@Pilchie 2021-09-20
Approved for EFCore 6.